pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "macds"
version = "1.0.0"
description = "Multi-Agent Coding Development System - A production-ready multi-agent system for automated software engineering"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "MACDS Contributors"}
]
keywords = [
    "multi-agent",
    "software-engineering",
    "automation",
    "llm",
    "ai-agents",
    "code-generation"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.10"
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "flake8>=6.1.0",
]
llm = [
    "openai>=1.0.0",
    "anthropic>=0.7.0",
]
all = [
    "macds[dev,llm]",
]

[project.scripts]
macds = "macds.main:main"

[project.urls]
Homepage = "https://github.com/macds/macds"
Documentation = "https://github.com/macds/macds/tree/main/docs"
Repository = "https://github.com/macds/macds"
Issues = "https://github.com/macds/macds/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["macds*"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["macds/tests"]
addopts = "-v --cov=macds --cov-report=term-missing"
